Refactor packages for entities and imports for it.
