+++ app/src/main/java/com/hfad/palamarchuksuperapp/data/services/FakeStoreApi.kt
@@ -1,7 +1,7 @@
 package com.hfad.palamarchuksuperapp.data.services
 
-import com.hfad.palamarchuksuperapp.data.entities.Product
-import com.hfad.palamarchuksuperapp.ui.common.ProductDomainRW
+import com.hfad.palamarchuksuperapp.data.dtos.ProductDTO
+import com.hfad.palamarchuksuperapp.domain.models.Product
 
 interface FakeStoreApi {
     suspend fun fetchProducts(): List<Product>
