+++ app/src/main/java/com/hfad/palamarchuksuperapp/data/services/GroqApi.kt
@@ -1,6 +1,6 @@
 package com.hfad.palamarchuksuperapp.data.services
 
-import com.hfad.palamarchuksuperapp.data.entities.AiModel
+import com.hfad.palamarchuksuperapp.data.dtos.GroqModelDTO
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
