+++ app/src/main/java/com/hfad/palamarchuksuperapp/domain/models/AiModel.kt
@@ -1,4 +1,4 @@
-package com.hfad.palamarchuksuperapp.data.entities
+package com.hfad.palamarchuksuperapp.domain.models
 
 import kotlinx.serialization.Required
 import kotlinx.serialization.SerialName
