+++ app/src/main/java/com/hfad/palamarchuksuperapp/domain/usecases/AddAiMessageUseCase.kt
@@ -1,6 +1,6 @@
 package com.hfad.palamarchuksuperapp.domain.usecases
 
-import com.hfad.palamarchuksuperapp.data.entities.MessageAI
+import com.hfad.palamarchuksuperapp.domain.models.MessageAI
 import com.hfad.palamarchuksuperapp.domain.repository.ChatAiRepository
 import kotlinx.coroutines.flow.update
 import javax.inject.Inject
