+++ app/src/main/java/com/hfad/palamarchuksuperapp/ui/compose/utils/BottomSheetSkill.kt
@@ -45,9 +45,7 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.hfad.palamarchuksuperapp.data.entities.Skill
-import com.hfad.palamarchuksuperapp.ui.common.SkillDomainRW
-import java.util.UUID
+import com.hfad.palamarchuksuperapp.domain.models.Skill
 
 
 @Suppress("detekt.FunctionNaming", "detekt.UnusedParameter", "detekt.LongMethod")
