+++ app/src/main/java/com/hfad/palamarchuksuperapp/ui/screens/BottomSheetFragment.kt
@@ -7,9 +7,8 @@ import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.google.android.material.datepicker.MaterialDatePicker
-import com.hfad.palamarchuksuperapp.data.entities.Skill
 import com.hfad.palamarchuksuperapp.databinding.ListItemBottomSheetBinding
-import com.hfad.palamarchuksuperapp.ui.common.SkillDomainRW
+import com.hfad.palamarchuksuperapp.domain.models.Skill
 import com.hfad.palamarchuksuperapp.ui.viewModels.SkillsChangeConst
 import com.hfad.palamarchuksuperapp.ui.viewModels.SkillsViewModel
 import java.text.SimpleDateFormat
